% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{assemble_key_pair}
\alias{assemble_key_pair}
\title{assemble_key_pair}
\usage{
assemble_key_pair(key, pair)
}
\arguments{
\item{key}{the key for your key pair}

\item{pair}{the pair for your key pair}
}
\value{
Returns your key pair if it exists or a blank string if it doesn't
exist
}
\description{
assemble_key_pair
}
\examples{
\dontrun{
limit <- NULL
limit <- assemble_key_pair('limit', limit)}
}
