% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{assemble_list}
\alias{assemble_list}
\title{assemble_config_object}
\usage{
assemble_list(character_vector)
}
\arguments{
\item{character_vector}{the character vector used to create the config object}
}
\value{
Returns your config object
}
\description{
assemble_config_object
}
\examples{
\dontrun{
limit <- NULL
limit <- assemble_key_pair('limit', limit)}
}
