% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_account_info}
\alias{get_account_info}
\title{get_account_info}
\usage{
get_account_info(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}

\item{address}{the address for which you're retrieving signatures}
}
\value{
Returns all information associated with the account of provided Pubkey
}
\description{
get_account_info
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_signature_for_address(url)}
}
