% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_block_height}
\alias{get_block_height}
\title{get_block_height}
\usage{
get_block_height(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the current block height of the node
}
\description{
get_block_height
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_block_height(url)}
}
