% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_genesis_hash}
\alias{get_genesis_hash}
\title{get_genesis_hash}
\usage{
get_genesis_hash(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the genesis hash
}
\description{
get_genesis_hash
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_genesis_hash(url)}
}
