% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_health}
\alias{get_health}
\title{get_health}
\usage{
get_health(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the current health of the node.
}
\description{
get_health
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_health(url)}
}
