% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_identity}
\alias{get_identity}
\title{get_identity}
\usage{
get_identity(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the identity pubkey for the current node
}
\description{
get_identity
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_identity(url)}
}
