% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_inflation_rate}
\alias{get_inflation_rate}
\title{get_inflation_rate}
\usage{
get_inflation_rate(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the specific inflation values for the current epoch
}
\description{
get_inflation_rate
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_inflation_rate(url)}
}
