% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_supply}
\alias{get_supply}
\title{get_supply}
\usage{
get_supply(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns information about the current supply.
}
\description{
get_supply
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_supply(url)}
}
