% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana_api_call.R
\name{get_version}
\alias{get_version}
\title{get_version}
\usage{
get_version(url, address)
}
\arguments{
\item{url}{the RPC url for your API call}
}
\value{
Returns the current Solana version running on the node
}
\description{
get_version
}
\examples{
\dontrun{
url <- "http://localhost:8899"
data <- get_version(url)}
}
